Create JES2_MSG_JOB_LOG
JES 2 job message job log.

Displays the job submitted from the JCL calling the COBOL program to execute into SDSF. Once the job runs, either it will fail or run successfully. In this message log, we have a Return Code RC=0000 , which means it ran successfully. 

This log also displays the Class on which it should run on A is Higher priority than B. In System Display and Search Facility (SDSF) is a utility that allows you to monitor, control, and view the output of jobs in the system. You can see on the left-hand side when the job was initiated with time and date along with the job number :

18.10.51 JOB08260 ---- WEDNESDAY, 30 NOV 2022 ----
18.10.51 JOB08260  IRR010I  USERID Z93595   IS ASSIGNED TO THIS JOB.

The Userid is permanently assigned with the execution of the job, whether it's an operator or a programmer. Ca7 or in SDSF

Along with the Step name and proc step the job had to process thru and the amount of time the execution of each step took place. 

Total time for the job to submit to process fully via each step was 0.10 MINUTES EXECUTION TIME